# docs/tasks-system.openapi.yaml
openapi: 3.1.0
info:
  title: Agentico — Tasks & Task I/O API
  version: "1.0.0"
  description: >
    REST surface for the Agentico Task System and the Task I/O / artifact-binding
    feature. The system spans **two services**:


    - **meetico** (`google-meet-express`) — owns the entire task domain: tasks,
      projects, columns, inputs, outputs, acceptance criteria, io_types,
      artifact_types, the I/O Review cockpit, the artifact-resolution /
      validation engine, and attestations. Mounted at `/tasks`.

    - **marketico** — owns the per-project **Notion** integration that the
      meetico `notion` resolver and the frontend pickers depend on (meetico never
      calls the Notion API directly; it proxies through marketico). Mounted at
      `/api/project-settings`.


    Data lives in Supabase schema **`ikigaigm`** (resolved per-tenant). Every I/O
    row carries a two-layer type: a semantic `io_type_id` (what the deliverable
    *is*) and a physical `artifact_type_id` (how it *resolves*), plus a `reference`
    locator. See `design-specs/tasks-system.md` for the conceptual model.
servers:
  - url: "{meeticoBase}/tasks"
    description: meetico — task domain (all `/tasks/**` operations, tag = meetico)
    variables:
      meeticoBase:
        default: https://localhost:5000
  - url: "{marketicoBase}/api/project-settings"
    description: marketico — Notion integration (all `/notion/**` operations, tag = marketico)
    variables:
      marketicoBase:
...
security:
  - bearerAuth: []
...
paths:
...
  /io-review/bind-preview:
    post:
      tags: [I/O Review]
      summary: Preview a binding (no persist)
      description: >
        Resolve a proposed binding live without persisting. Pass a physical
        `artifact_type_id` and a locator as either `url` (parsed into a
        reference by the type's parser) or a ready `reference` object.
      operationId: bindPreview
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BindRequest" }
      responses:
        "200":
          description: Normalized reference + live resolution
          content:
            application/json:
              schema:
                type: object
                properties:
                  reference: { $ref: "#/components/schemas/ArtifactReference" }
                  resolved: { $ref: "#/components/schemas/ResolvedArtifact" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        meetico `/tasks` routes use passport-jwt (`Authorization: Bearer <token>`),
        falling back to session. marketico `/notion` routes resolve a per-request
        tenant schema; a missing tenant yields HTTP 400.
...
  responses:
...
    NotFound:
      description: Resource not found.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
...
    ArtifactReference:
      type: object
      description: >
        Physical locator, shape depends on the artifact type's parser. Examples —
        gdrive `{ file_id, tab?, url? }`; notion `{ page_id, url?, project_id? }`;
        sql `{ query, params? }`; storage `{ bucket, path }`; api `{ url }`;
        inline `{ content }`.
      additionalProperties: true
...
    ResolvedArtifact:
      type: object
      description: Uniform shape returned by `resolveArtifact` for every resolver kind.
      properties:
        exists: { type: boolean }
        content_text: { type: string, nullable: true }
        url: { type: string, nullable: true }
        record: { nullable: true }
        metadata: { type: object, additionalProperties: true, nullable: true }
        error: { type: string, nullable: true }

    # ---- binding ----
    BindRequest:
      type: object
      required: [artifact_type_id]
      description: >
        Provide the locator as `url` (parsed by the type's parser) OR a ready
        `reference` object. `project_id` is required for notion references.
      properties:
        artifact_type_id: { type: string, format: uuid }
        url: { type: string }
        reference: { $ref: "#/components/schemas/ArtifactReference" }
        project_id: { type: string, format: uuid }

    # ---- I/O review ----
...